requirements.txt
flask
flask_sqlalchemy
pymysql
flask_wtf
wtforms
pytest
flask_testing

